Setup cbor attachment pipeline for index and use as default pipeline.
Setup the cbor-attachments pipeline:
1. Iterate over all attachments (field) and use the attachment processor
   to extract text data from the binary/data field called 'data'.
2. Iterate over all extracted data and extend the SearchableText.
   At same time remove the binary data to save some space. With
   Version 8 of elasticsearch we could use the remove_binary flag, but
   it does not exist yet in version 7.
It's called cbor-attachment pipeline, because the data is transfered
using cbor. This eliminates the base64 overhead.
Also the cbor-attachments pipline is used as default pipeline for the index.
There is a condition defined, so the pipeline is only triggered if
actual binary data is available to extract.